requirements.txt
python-telegram-bot
openai-whisper
pydub
moviepy
python-docx
